backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
sqlalchemy
google-generativeai
openai
